fix(seguridad): acota offsetBase para cerrar el robo de pocket por desbordamiento
El indice de punto del REQ_SRA_UNLOCK_CHAIN llegaba del wire con un
Integer.parseInt a pelo, sin rango, y el handler calculaba el desplazamiento
como pointIdx * 32 en aritmetica int. Como 2^27 puntos son exactamente 2^32
bytes, un offsetBase = mySlot*2 + 2^27 vuelve a caer sobre el MISMO punto del
megapacket tras el wraparound, pero deja de coincidir con las igualdades del
guard que protege el pocket propio, y el guard de rango tampoco salta porque
(pointIdx+1)*32 tambien se desborda a un valor pequeno. Un host hostil podia
pedirle a un peer que pelara su propio candado sobre su propia carta y leerle
la mano, justo lo que el comentario del handler dice impedir.

Se cierra en dos capas. UnlockChainWire acota offsetBase (y offsetBase +
numChains) a MAX_POINT_INDEX al parsear, asi que un REQ con un indice absurdo
se rechaza entero antes de llegar al handler. Y el handler pasa a long para el
calculo del offset y la comprobacion de rango, de modo que sus guards operan
sobre valores exactos aunque el corte de arriba no estuviera.

El test del ataque (PocketSelfStripAttackTest) gana la variante desbordada:
demuestra que el offset con 2^27 apunta al mismo byte, que el guard escrito con
igualdades no lo reconoce, y exige que el wire lo rechace. Sin el fix falla al
parsear el REQ. Los 20 tests de crypto existentes siguen pasando.

+++ src/main/java/com/tonikelope/coronapoker/WaitingRoomFrame.java
@@ -2826,8 +2826,14 @@ public class WaitingRoomFrame extends JFrame {
                                                                         }
                                                                         java.util.List<String> outChains = new java.util.ArrayList<>();
                                                                         for (int j = 0; j < it.chains.size(); j++) {
-                                                                            int pointIdx = it.offsetBase + j;
-                                                                            if (pointIdx < 0 || (pointIdx + 1) * 32 > megapacket.length) {
+                                                                            // Aritmetica en long a proposito: offsetBase viene del wire y
+                                                                            // pointIdx * 32 en int se desborda con 2^27 puntos (= 2^32
+                                                                            // bytes), cayendo de nuevo dentro del rango valido y burlando
+                                                                            // a la vez este guard y las igualdades de slot de mas abajo.
+                                                                            // UnlockChainWire ya acota offsetBase al parsear; esto lo cierra
+                                                                            // tambien aqui, sin depender de aquello.
+                                                                            long pointIdx = (long) it.offsetBase + j;
+                                                                            if (pointIdx < 0 || (pointIdx + 1) * 32L > megapacket.length) {
                                                                                 LOGGER.log(Level.SEVERE, "ZERO-TRUST: REQ_SRA_UNLOCK_CHAIN offset out of range — refusing");
                                                                                 crupier.triggerSecurityLockdown(Translator.translate("zero_trust.host_bad_wire"));
                                                                                 return;
@@ -2859,7 +2865,7 @@ public class WaitingRoomFrame extends JFrame {
                                                                                 crupier.triggerSecurityLockdown(Translator.translate("zero_trust.host_pocket_extraction"));
                                                                                 return;
                                                                             }
-                                                                            byte[] point = java.util.Arrays.copyOfRange(megapacket, pointIdx * 32, (pointIdx + 1) * 32);
+                                                                            byte[] point = java.util.Arrays.copyOfRange(megapacket, (int) (pointIdx * 32L), (int) ((pointIdx + 1) * 32L));
                                                                             DealChain.Extended ext = DealChain.extend(point, it.chains.get(j), commitments, local_nick, myLock);
                                                                             if (ext == null) {
                                                                                 LOGGER.log(Level.SEVERE, "ZERO-TRUST: REQ_SRA_UNLOCK_CHAIN chain not anchored/invalid (offset {0}) — extraction or tampering, refusing", pointIdx);
